test: implement simple perf tests
This patch introduces simple performance tests for select and
insert. Tests are really simple and could catch only huge
regressions. The purpose to merge it into master to have simple
way to check performance and such tests shouldn't be broken if
they work in CI.
It worth to note that CI is public and tests run in containers that
mean that result could vary from run to run.