introduce internal API for raw tuples
This patch introduces internal API for insert, replace and upsert
raw tuples. To reduce amonut of changes this methods will be
exposed in next patches. Hovewer this patch already tries to
support naming convenstion: we internally use module.object for
operations on objects and module.tuple for operations on tuples.

Also some kind of optimization is introduced. Before this patch we
call unflatten for object twice. This patch removes one of them.
This way reduce amount of Lua garbage.

Part of #6