replace "sharding key" to "bucket_id" in error message
Initially we expect that first field of "bucket_id" index is
empty as user shouldn't affect internal processes in our system.
So we return an error in case then this field in tuple is not
empty.

But some logic error in error message was done. When we say
"sharing key" we mean a part of tuple (currently primary key for
bucket_id calculation). May be "shard key" is more vshard-agnostic
varian. So to avoid over complication let's use "bucket_id" as in
vshard.