ci: show rocks server hostname in logs
There is no reason to hide a server that will serve the rock out. Now it
is always rocks.tarantool.org, but it may be not so obvious for one, who
don't know our infrastructure details.

If we'll have more than one rocks server in a future (say, for testing
purposes), the explicit server hostname in the CI logs may be helpful
for debugging infrastructural problems.

See also [1].

[1]: https://github.com/tarantool/rocks.tarantool.org/issues/22