Support partial results (#118)
User may need only a subset of the query fields, but receives the values ​​of all fields, which leads to unnecessary network load. The solution is fields filtration on storage side by user-specified option. `fields` option allows to get only a subset of fields. This parameter contains the names of the fields that the user requests.