Use "merger" for fetch&sort tuples on router (#38)
Before this patch we used hand-crafted heap to sort tuples on
router. This patch removes it and changes to "merger" module
written in C.

"Merger" is available since Tarantool 2.2. Later "tuple-merger"
was introduced it's a version that is compatible with some
Tarantool 1.10 versions. In fact support several Tarantool
versions is quite complex task. Some msgpack helpers is not
avaliable in Tarantool 1.10 to solve this problem small
compatibility lawer is implemented.
Also "tuple-keydef" doesn't support collation_id options and it's
not so for built-in "key_def" module. We need to normalize index
parts and change "collation_id" to "collation" when create
"key_def" instance.

For versions where merger is not supported/available yet
(before 1.10.8 and before 2.3.3) we use previous approach with
lua-implemented heap.

crud.select.compat.select - default module with merged
crud.select.compat.select_old - compatibility module without 
merger support written in Lua

Co-authored-by: AnaNek <neknas28@gmail.com>