test: make simple_operations_test compatible with Tarantool 2.8
After https://github.com/tarantool/tarantool/issues/3378 it's
possible to perform updates with gaps. So one of testcases that
returned an error became irrelevant for 2.8+. This patch checks
that update is valid for new Tarantool versions.

Also it's not complete 2.8 support because of a bug
https://github.com/tarantool/tarantool/issues/6069. But since
issue has been already resolved we should just wait for update.