Fix pagination + multipart indexes (#177)
Fixed invalid results for pagination queries when filtering on
a parts of a composite index. Now, instead of "scrolling" 
to the desired tuple (during pagination) using the 
``scroll_to_after_tuple`` function, we change the iterators 
(``EQ`` -> ``GE``, ``REQ`` -> ``LT``) to use them in 
``index:pairs`` function, together with a combination with 
additional filtration. Closes #170