drop heap module
This is really happy case because currently we can drop part of
crud module (heap) and replace it with a heap that shipped with
vshard.

Interfaces is not the same but it's easy to fix. Also I want to
note that heap is used for an ancient 1.10 version (that didn't
support tuple-keydef/tuple-merger modules) and 2.2 that has merger
that segfaults. There is no significant benefits to keep our heap
implementation if we could use heap directly from vshard.