Support custom sharding key in select
NOTE: Prior to this patch CRUD assumes an index is unique. It was true
for the primary key, but it is not guaranteed for a sharding key. Patch
adds a tests with select() for non-unique index that failed due to
assumption regarding uniq index in crud/select/plan.lua. Seems we
can remove this condition and fix tests that relies on
total_tuple_count == 1. See also related discussion in [1].

1. https://github.com/tarantool/crud/pull/181#discussion_r707761937

Part of #166

Reviewed-by: Oleg Babin <babinoleg@mail.ru>
Reviewed-by: Alexander Turenko <alexander.turenko@tarantool.org>