test: allow to count select requests on storages
I don't want to lean on box.stat() information here, because I don't
control all iproto calls to storages: say, vshard rebalancer may perform
them in background.

Instead, I wrapped particular storage function I'm interested in.

The goal is to be able to determine how much storages are involved into
a select/pairs request.

It is implemented as a helper for testing, but hopefully we'll implement
some nice statistics as part of the module in a future (see #224).

Part of #220