ci: fix testing with version 1.10.6
We need a version 1.10.6 in testing because it has no tuple-keydef and
tuple-merger support and we can test our own implementation of both
modules on that version. However Tarantool EE 1.10.6 is not available
for installation anymore, so we will use Tarantool CE 1.10.6 as a
replacement.

Luarocks in Tarantool 1.10.6 has a bug [1], so we need to apply a patch
[2] with fix for tarantool's luarocks file before using `tarantoolctl
rock`.

1. https://github.com/tarantool/tarantool/issues/4704
2. https://github.com/tarantool/luarocks/commit/eb555fa4073f4dc8278b68b04725ef5de2b54f0a