Fix merger support recognition on tarantool 2.10+
We cannot use just string comparison as before, because '2.10.0-<...>'
less than, say, '2.3' lexicographically. The module had to use the old
select implementation on tarantool 2.10 due to this problem.

Implemented more accurate checks, whether built-in and external
tuple-merger are supported. Say, it'll not enable the built-in merger on
tarantool 2.4.1, where the module has a known problem (see comments in
the code).

Fixed the problem of the same kind in a test case.

When I replaced `_G._TARANTOOL` with just `_TARANTOOL`, the luacheck
linter said me:

 | <...>/test/helper.lua:307:31: accessing undefined field match of
 | global _TARANTOOL

On the line:

 | local major_minor_patch = _TARANTOOL:split('-', 1)[1]

But when I added '_TARANTOOL' into globals, it stops to complain.
Strange, but okay. I prefer just `_TARANTOOL` without `_G`.