utils: move a couple of feature checkers to test/
There are several checkers of a feature presence in tarantool, which are
used only in tests. Moved them to the testing code.

I preferred code simplicity over a negligible speedup in tests, so left
re-parsing of tarantool version for each call to a checker.

Clarified behaviour of `helpers.tarantool_version_at_least()` and marked
it for future removal: it is easy to use it improperly.