readme: fix update_cache() recipe
When a sharding key is updated on storages a user should call this
function on routers to re-fetch the new sharding keys. I would highlight
that even if we add a space that was never seen before, we should call
the function on routers. Otherwise crud will assume that the new space
has the sharding key equal to the primary key and will calculate
`bucket_id` incorrectly.

Follows up #166
Related to #212
Related to TNT-462