ci: workaround luacov-coveralls installation fail
GitHub had disabled insecure git clone (see [1]), so we should use
`git+https://` in the rockspec instead of `git://`. The fix is proposed
to the upstream in [2], but is not merged at the moment of writting
this. This patch workarounds the problem on our side.

Recent luarocks (3.8.0) autoreplaces `git://` with `git+https://`, but
we have older luarocks version shipped with tarantool. See [3] for
details.

[1]: https://github.blog/2021-09-01-improving-git-protocol-security-github/
[2]: https://github.com/moteus/luacov-coveralls/pull/30
[3]: https://github.com/tarantool/tarantool/issues/6597