Add truncate space on cluster in tests for stats
Before this commit `simple_operation_cases`
were organized as map
(table indexed not with numbers),
in Lua iteration over map does not occur in
the order in which the elements were specified
in the map.
But simple operation tests could fail
in case if tests would be executed not
in the order in which they are specified,
because, for example, if `replace()` is performed
before `insert()`, an error will be received.
So simple operation tests are codependent.
To solve this problem `truncate_space_on_cluster`
was added after each simple operation test.

Part of #193