select: fix EQ with negative first results
EQ with after has always been replaced by LT. The idea was to start
traverse not from the after value, but from a next value. The
micro-optimization allowed to fetch one less tuple per time.

It worked well for unique indexes, but this led to wrong results for
non-unique indexes.