select: fix LE and after results when the condition value < after
LE with after has always been replaced by LT. The idea was to start
traverse not from the after value, but from a next value. The
micro-optimization allowed to fetch one less tuple per time.

But this caused the condition value to be skipped if the value is less
than the after value.