ddl: fix fetching invalid configuration
Before this patch, in case there is a sharding key record
in _ddl_sharding_key for a non-existing space (for example, if ddl space
wasn't properly dropped) and sharding info needs to be updated, any crud
request will fail with "attempt to index a nil value" error. This patch
fixes the behavior and adds a log warning if ddl configuration is
invalid.

Closes #308