fix: pre-hotreload cartridge support
`crud` module is cartridge-independent in nature, but provides cartridge
roles which are the most popular way to setup the module. The roles also
not use any modern cartridge features and should work with any cartridge
version. But since crud.cfg was introduced [1], it was required to add
some code for roles reload [2] proper support. Now
cartridge.hotreload module is unconditionally required, so roles cannot
be used with cartridge older than 2.4.0. This patch fixes the behavior.

1. https://github.com/tarantool/crud/commit/6da4f5684a00fe39106ac139538b06821c8c829c
2. https://github.com/tarantool/cartridge/commit/941952e004c86c5883f492a63f62d98d2b3453af

Follows #244