coverage: fix loaded function name
Now we load flatten functions for spaces with `@flatten` source. Source
starting with @ assumes that it describes the file name, yet here it
isn't. Then luacov fails to find "flatten" file on coverage info build.

When we do not specify source explicitly, the whole string became the
source. It will not confuse luacov anymore (it just ignores them), and
it's more helpful for debug.

1. https://www.lua.org/manual/5.2/manual.html#pdf-load

Closes #249