schema: fix stale schema for metadata generation
Corrects using of an stale schema to generate metadata for
`crud.update`, `crud.insert`, `crud.insert_*`, `crud.replace`,
`crud.replace_*`, `crud.upsert`, `crud.upsert_*`, `crud.delete`,
`crud.max`, `crud.min`, `crud.select`, `crud.pairs` and `crud.get`.

If the implemented `fetch_latest_metadata` option is used,
it is guaranteed that the metadata will be up-to-date.
Before receiving the space format, a mismatch check will be
performed between the scheme version on all involved storage
and the scheme version in the net_box connection of the router.
In case of mismatch, the schema reload will be triggered.

Closes #236