test: support tests with vshard
The patch updates all tests to work with `vshard` and `cartridge`
at the same time. But tests of the `cartridge` role of the cartridge
or special features are skipping for `vshard`.

If `cartridge` is not installed or not supported (Tarantool 3.0+),
tests will run only with `vshard`.

The source of `vshard_helpers` folder is `vshard` repository [1] with
minor modifications.

1. https://github.com/tarantool/vshard/tree/b3c27b32637863e9a03503e641bb7c8c69779a00/test/luatest_helpers

Closes #364