rock: drop external tuple merger and tuple keydef installation
Crud requires a tuple merger to work, and it currently works with three
kinds of mergers: an embedded Tarantool one, an external one, and a crud
embedded one. The latter implementation is suboptimal but does not depend
on any additional symbols.

The external merger replicates Tarantool's embedded merger, and serves two
purposes:
1. Bring tuple merging to Tarantool 1.10 which does not have an embedded
   merger.
2. Allows for versioning the merger module independently from Tarantool.

Because of potential symbol dependency conflicts, the external merger is
installed in a hacky way, which didn't actually work when crud is built as
part of another rockspec.

Since Tarantool 1.10 can work with the merger embedded to crud, there is no
need in pulling an external merger, so let's drop this dependency.

Instead, add a CI matrix case with external merger installation to test
against it.

All written above also applies to the tuple keydef module.

Needed for tarantool/tarantool#8147