crud: allow tuple fields of select options to be cdata in type checks
In scope of tarantool/tarantool#8147, box tuples returned from remote
procedure calls our now encoded as a new extension, `MP_TUPLE`, and, as,
the results of such calls are now decoded as box tuples, i.e., cdata. For
compatibility with this feature, we need to allow tuple fields of select
options, which are returned from a remote procedure calls, to be cdata in
option type checks.

Needed for tarantool/tarantool#8147