test: fix flaky update unflatten case
The following sequence of events happens:
- data inserted with crud,
- schema updated 12 times,
- data updated with crud.

For Tarantool supporting field name in updates, there is no additional
check for a field name presenting in current format. So it is possible
to get outdated metadata.

For some reason, new identification mode had uncovered this issue. It
is likely related to the change in the order of servers in maps.