crud: support vshard with master discovery
In vshard 0.1.25 and Tarantool 3.0 new feature has been implemented.
The feature is related to automatic master discovery. If it has been
enabled, crud fails to bootstrap and work in several places.
This feature is enabled by Tarantool 3.0 if vshard cluster was
configured with specific 3.0 config flags.

Unfortunately, it is impossible to rework the code using only
vshard public API now [2].

1. https://github.com/tarantool/vshard/issues/429
2. https://github.com/tarantool/vshard/issues/467

Closes #409