test: call router for router operations
Before this patch, most tests used an assumption that main server is
a router. Yet these are two different concepts.

`main_server` entity was introduced in Cartridge test helpers Cluster
as some server you may refer to for maintenance work, like clusterwide
configuration update. Any cluster server is fine to do such operation:
router or not. Since most cluster configurations start with router
definition, it is often a first server in a list and is chosen to be
main_server as well. But no one guarantees that main_server provides
router API (both for vshard and crud).

Part of #412
Part of #415