test: wait for schema bootstrap
Tarantool 3.0 instances start in ro mode. It is possible that masters
are not available for write requests right after test case has been
started.

This patch supports only vshard tests for Tarantool 2.2+. After this
patch, schema init is async and we wait for schema to initialize in
tests.

Part of #412
Part of #415