test: support tarantool 3 cluster from config
After this patch, it is possible to start a Tarantool 3.0 cluster
from the config for tests. It doesn't yet integrated with common test
matrix yet. The tests are basic and shallow since it would be tested
against the whole crud test suite in the next commits of the patchset.

Part of #415