test: remove vinyl from matrix
This patch removes vinyl from test matrices. Since [1], vinyl is
considered experimental and we do not guarantee its work with modules.
Vinyl clusters like to misbehave in tests, often resulting in failing
tests in CI. To improve developer experience, this patch disables them.

This patch do not remove vinyl support since there is no vinyl-specific
code in crud, but we do not guarantee it anymore.

1. https://github.com/tarantool/doc/issues/3523

Part of #412
Part of #415