call: support vshard's request_timeout in read calls
vshard 0.1.28 introduces a new option for read requests:
`request_timeout`.

`request_timeout` serves as a protection against hung replicas. When it
is specified, it must be <= `timeout`, and it introduces new behavior
for router calls: they start retrying TimedOut errors, each time trying
the request on a new replica, as long as `timeout` (the old one)
permits.

Support the option only in crud requests with `mode = 'read'`, since
it makes no sense with `mode = 'write'`. There is no one to retry the
request on if master is unresponsive.